requirements.txt
pandas
numpy
scikit-learn
streamlitgit 